Add flag DSV3_USE_FP8_GEMM (#10133)
* add flag DSV3_USE_FP8_GEMM

* fix
